fix if conditions in registration flow
